后台保存借款人信息BUG修复；currentUser增加借款人ID缓存；其他BUG修复

+++ com.jebao.root/3.1-com.jebao.jebaodb.dao/src/main/java/com/jebao/jebaodb/dao/mapper/loaner/TbLoanerMapper.java
@@ -7,9 +7,9 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface TbLoanerMapper {
-    int insert(TbLoaner record);
+    Long insert(TbLoaner record);
 
-    int insertSelective(TbLoaner record);
+    Long insertSelective(TbLoaner record);
 
     TbLoaner selectByPrimaryKey(Long lId);
